requirements.txt
guizero
paho-mqtt
